+++ lib/packer/builder.rb
@@ -1,17 +1,4 @@
 # Encoding: utf-8
-# Copyright 2014 Ian Chesal
-#
-#    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compliance with the License.
-#    You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#    See the License for the specific language governing permissions and
-#    limitations under the License.
 require 'packer/builders/all'
 
 module Packer
@@ -20,12 +7,14 @@ module Packer
     AMAZON_INSTANCE = 'amazon-instance'
     DOCKER          = 'docker'
     VIRTUALBOX_ISO  = 'virtualbox-iso'
+    NULL            = 'null'
 
     VALID_BUILDER_TYPES = [
       AMAZON_EBS,
       AMAZON_INSTANCE,
       DOCKER,
-      VIRTUALBOX_ISO
+      VIRTUALBOX_ISO,
+      NULL
     ]
 
     class UnrecognizedBuilderTypeError < StandardError
@@ -39,7 +28,8 @@ module Packer
         AMAZON_EBS      => Packer::Builder::Amazon::EBS,
         AMAZON_INSTANCE => Packer::Builder::Amazon::Instance,
         DOCKER          => Packer::Builder::Docker,
-        VIRTUALBOX_ISO  => Packer::Builder::VirtualBoxISO
+        VIRTUALBOX_ISO  => Packer::Builder::VirtualBoxISO,
+        NULL            => Packer::Builder::Null
       }.fetch(type).new
     end