+++ lib/packer/provisioner.rb
@@ -1,28 +1,27 @@
 # Encoding: utf-8
-# Copyright 2014 Ian Chesal
-#
-#    Licensed under the Apache License, Version 2.0 (the "License");
-#    you may not use this file except in compliance with the License.
-#    You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS,
-#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#    See the License for the specific language governing permissions and
-#    limitations under the License.
 require 'packer/provisioners/all'
 require 'packer/dataobject'
 
 module Packer
   class Provisioner < Packer::DataObject
-    SHELL = 'shell'
-    FILE = 'file'
+    SHELL             = 'shell'
+    FILE              = 'file'
+    SALT              = 'salt-masterless'
+    ANSIBLE           = 'ansible-local'
+    CHEF_CLIENT       = 'chef-client'
+    CHEF_SOLO         = 'chef-solo'
+    PUPPET_MASTERLESS = 'puppet-masterless'
+    PUPPET_SERVER     = 'puppet-server'
 
     VALID_PROVISIONER_TYPES = [
       SHELL,
-      FILE
+      FILE,
+      SALT,
+      ANSIBLE,
+      CHEF_CLIENT,
+      CHEF_SOLO,
+      PUPPET_MASTERLESS,
+      PUPPET_SERVER
     ]
 
     class UnrecognizedProvisionerTypeError < StandardError
@@ -33,8 +32,14 @@ module Packer
         raise UnrecognizedProvisionerTypeError.new("Unrecognized provisioner type #{type}")
       end
       {
-        SHELL => Packer::Provisioner::Shell,
-        FILE  => Packer::Provisioner::File,
+        SHELL             => Packer::Provisioner::Shell,
+        FILE              => Packer::Provisioner::File,
+        SALT              => Packer::Provisioner::Salt,
+        ANSIBLE           => Packer::Provisioner::Ansible,
+        CHEF_CLIENT       => Packer::Provisioner::Chef::Client,
+        CHEF_SOLO         => Packer::Provisioner::Chef::Solo,
+        PUPPET_MASTERLESS => Packer::Provisioner::Puppet::Masterless,
+        PUPPET_SERVER     => Packer::Provisioner::Puppet::Server
       }.fetch(type).new
     end